eduGAIN import: does not set mdrpi:*
The metadata parser worked under the assumption that all relevant
metadata is within the IDPSSODescriptor or SPSSODescriptor.

This is not the case as it possible to have a Extensions element
before IDPSSODescriptor or SPSSODescriptor. We therefore need to parse
the Extensions element and store the intermediate results and add them
- if present - to the actual metaData of the result.